split sql block based tests into their own file
