+++ packages/malloy/src/lang/test/parse.spec.ts
@@ -27,17 +27,14 @@ import {
   DocumentPosition,
   SQLBlockSource,
   SQLBlockStructDef,
-  StructDef,
   expressionIsCalculation,
   isFieldTypeDef,
   isFilteredAliasedName,
   isSQLFragment,
 } from '../../model';
-import {makeSQLBlock} from '../../model/sql_block';
 import {
   MarkedSource,
   TestTranslator,
-  aTableDef,
   getExplore,
   getField,
   getJoinField,
@@ -49,17 +46,6 @@ import {
 import isEqual from 'lodash/isEqual';
 import {isGranularResult} from '../ast/types/granular-result';
 import './parse-expects';
-
-function unlocatedStructDef(sd: StructDef): StructDef {
-  const ret = {...sd};
-  ret.fields = sd.fields.map(f => {
-    const nf = {...f};
-    delete nf.location;
-    return nf;
-  });
-  delete ret.location;
-  return ret;
-}
 
 type TestFunc = () => undefined;
 
@@ -1076,130 +1062,6 @@ describe('unspported fields in schema', () => {
       'source: x is a { dimension: notUn is aun::string }'
     );
     expect(uModel).modelCompiled();
-  });
-});
-
-describe('sql:', () => {
-  function makeSchemaResponse(sql: SQLBlockSource): SQLBlockStructDef {
-    const cname = sql.connection || 'bigquery';
-    return {
-      type: 'struct',
-      name: sql.name,
-      dialect: "standardsql'",
-      structSource: {
-        type: 'sql',
-        method: 'subquery',
-        sqlBlock: {
-          type: 'sqlBlock',
-          ...sql,
-          selectStr: sql.select.filter(s => typeof s === 'string').join(''),
-        },
-      },
-      structRelationship: {type: 'basetable', connectionName: cname},
-      fields: aTableDef.fields,
-    };
-  }
-  test('definition', () => {
-    const selStmt = 'SELECT * FROM aTable';
-    const model = new TestTranslator(`
-      sql: users IS {
-        select: """${selStmt}"""
-        connection: "aConnection"
-      }
-    `);
-    const needReq = model.translate();
-    expect(model).modelParsed();
-    const needs = needReq?.compileSQL;
-    expect(needs).toBeDefined();
-    if (needs) {
-      const sql = makeSQLBlock([{sql: selStmt}], 'aConnection');
-      expect(needs).toMatchObject(sql);
-      const refKey = needs.name;
-      expect(refKey).toBeDefined();
-      if (refKey) {
-        const sr = makeSchemaResponse(sql);
-        model.update({compileSQL: {[refKey]: sr}});
-        expect(model).modelCompiled();
-        expect(unlocatedStructDef(model.sqlBlocks[0])).toEqual(
-          unlocatedStructDef({...sr, as: 'users'})
-        );
-      }
-    }
-  });
-  test('source from sql', () => {
-    const selStmt = 'SELECT * FROM aTable';
-    const model = new TestTranslator(`
-      sql: users IS { select: """${selStmt}""" }
-      source: malloyUsers is from_sql(users) { primary_key: ai }
-    `);
-    expect(model).modelParsed();
-    const needReq = model.translate();
-    const needs = needReq?.compileSQL;
-    expect(needs).toBeDefined();
-    if (needs) {
-      const sql = makeSQLBlock([{sql: selStmt}], 'aConnection');
-      const refKey = needs.name;
-      model.update({compileSQL: {[refKey]: makeSchemaResponse(sql)}});
-      expect(model).modelCompiled();
-      const users = model.getSourceDef('malloyUsers');
-      expect(users).toBeDefined();
-    }
-  });
-  test('explore from imported sql-based-source', () => {
-    const selStmt = 'SELECT * FROM aTable';
-    const createModel = `
-      sql: users IS { select: """${selStmt}""" }
-      source: malloyUsers is from_sql(users) { primary_key: ai }
-    `;
-    const model = new TestTranslator(`
-      import "createModel.malloy"
-      source: importUsers is malloyUsers
-    `);
-    model.importZone.define(
-      'internal://test/langtests/createModel.malloy',
-      createModel
-    );
-    expect(model).modelParsed();
-    const needReq = model.translate();
-    const needs = needReq?.compileSQL;
-    expect(needs).toBeDefined();
-    const sql = makeSQLBlock([{sql: selStmt}]);
-    model.update({compileSQL: {[sql.name]: makeSchemaResponse(sql)}});
-    expect(model).modelCompiled();
-  });
-  it('turducken', () => {
-    const m = new TestTranslator(`
-      sql: someSql is {
-        select: """SELECT * FROM %{ a -> { group_by: astr } }% WHERE 1=1"""
-      }
-    `);
-    expect(m).modelParsed();
-    const compileSql = m.translate().compileSQL;
-    expect(compileSql).toBeDefined();
-    if (compileSql) {
-      const select = compileSql.select[0];
-      const star = compileSql.select[1];
-      const where = compileSql.select[2];
-      expect(select).toEqual({sql: 'SELECT * FROM '});
-      expect(isSQLFragment(star)).toBeFalsy();
-      expect(where).toEqual({sql: ' WHERE 1=1'});
-    }
-  });
-  it('model preserved', () => {
-    const selStmt = 'SELECT * FROM aTable';
-    const shouldBeOK = `
-      source: newa is a
-      sql: someSql is { select: """${selStmt}""" }
-      source: newaa is newa
-    `;
-    const model = new TestTranslator(shouldBeOK);
-    expect(model).modelParsed();
-    const needReq = model.translate();
-    const needs = needReq?.compileSQL;
-    expect(needs).toBeDefined();
-    const sql = makeSQLBlock([{sql: selStmt}]);
-    model.update({compileSQL: {[sql.name]: makeSchemaResponse(sql)}});
-    expect(model).modelCompiled();
   });
 });
